chore: Removing IViewHostProvider as not required
