--- src/Uno.Extensions.Navigation.UI/IViewHostProvider.cs
@@ -1,6 +1,0 @@
-﻿namespace Uno.Extensions.Navigation.UI;
-
-internal interface IViewHostProvider
-{
-	void InitializeViewHost(FrameworkElement contentControl, Task InitialNavigation);
-}
